DinoCode++
The expression evaluator is more present within the general syntax. Calling methods or values within objects with the traditional dot syntax (OBJECT.object.value) is now supported both as a parameter and as a complete statement.

Value assignment now supports object references, it is possible to assign values within a complex content tree. (Set OBJECT.object.value = "NEW VALUE")

It is now possible to call native functions with their traditional syntax: FunctionName(Param, Param2, Param3, ParamN). This does NOT apply to DinoCode TAGs.

The expansion pointer for Arrays/Objects can now be used even at the end of Expressions that will produce an object: $(Object)*

Some additional AHK variables like ComSpec, ErrorLevel, ProgramFiles are now accessible.

Added special support to the "Clipboard" Variable. It is a functional Variable only within a value assignment, since the assigned value will completely replace the user's Clipboard. "Clipboard" cannot be recovered as a normal variable, attempting to use it outside of an assignment will break the script.

Changed the reference marker for expression escaping, this reduces the chances of conflicts during the interpretation of strings with unusual characters.

Fixed a potential bug when calling TAGs asynchronously when there is a loop in process.

It is now possible to use "continue" inside a loop to skip to the next iteration.

New isNumber function (integers, decimals)